fixing errors

+++ src/job_position/application/commands/delete_job_position.py
@@ -1,5 +1,7 @@
 from dataclasses import dataclass
 
+from src.candidate_application.domain.repositories.candidate_application_repository_interface import \
+    CandidateApplicationRepositoryInterface
 from src.job_position.domain.exceptions import JobPositionNotFoundException
 from src.job_position.domain.value_objects import JobPositionId
 from src.job_position.infrastructure.repositories.job_position_repository import JobPositionRepositoryInterface
@@ -12,12 +14,24 @@ class DeleteJobPositionCommand(Command):
 
 
 class DeleteJobPositionCommandHandler:
-    def __init__(self, job_position_repository: JobPositionRepositoryInterface):
+    def __init__(self,
+                 job_position_repository: JobPositionRepositoryInterface,
+                 candidate_application_repository: CandidateApplicationRepositoryInterface,):
         self.job_position_repository = job_position_repository
+        self.candidate_application_repository = candidate_application_repository
+
+
 
     def execute(self, command: DeleteJobPositionCommand) -> None:
         job_position = self.job_position_repository.get_by_id(command.id)
+
         if not job_position:
             raise JobPositionNotFoundException(f"Job position with id {command.id.value} not found")
 
-        self.job_position_repository.delete(command.id)
+        candidates = self.candidate_application_repository.get_applications_by_position(command.id)
+
+        if candidates:
+            job_position.close_position()
+            self.job_position_repository.save(job_position)
+        else:
+            self.job_position_repository.delete(command.id)
